move cmd/everything to top level directories
